Move decode_image_message to sobits_vla_common; deploy no longer uses cv_bridge
